--- src/InterviewGenerator.Application/Interfaces/ILoginService.cs
@@ -1,11 +1,0 @@
-﻿using InterviewGenerator.Application.Dto;
-using InterviewGenerator.Application.ViewModels;
-using InterviewGenerator.Domain.Entidade.Common;
-
-namespace InterviewGenerator.Application.Interfaces
-{
-    public interface ILoginService
-    {
-        Task<ResponseBase<LoginViewModel>> BuscarTokenUsuario(GerarTokenUsuarioDto usuario);
-    }
-}
